convert inactive pools to Stack>GameObject>, convert active pools to single Hashtable, add uint pool identifier to Poolable for Hashtable key, set type identifier variables in enemies to -2 to ease debugging and prevent infinite spawning of standard enemies, move splitter child spawning into for loop

+++ Hive Armada/Assets/Scripts/Enemies/JumperTurret.cs
@@ -24,7 +24,7 @@ namespace Hive.Armada.Enemies
         /// <summary>
         /// Type identifier for this enemy's projectiles in objectPoolManager
         /// </summary>
-        private int projectileTypeIdentifier;
+        private short projectileTypeIdentifier;
 
         /// <summary>
         /// The point where this enemy shoots from.
